Simplified SQLite and Redis merge readers
